+++ src/main/java/br/com/milvusartis/ecommerce/model/dto/PedidoItemDTO.java
@@ -5,8 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -16,6 +14,6 @@ public class PedidoItemDTO {
     private Long idPedido;
     private Produto idProduto;
     private Integer nrQuantidade;
-    private BigDecimal vlProduto;
+    private Double vlProduto;
 
 }
